refactor: moved all map search functionality to home cubit, remove dummy test widgets, cleaned up search ui
